+++ errors/CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## To be Released
 
-* feat: add New method to `ErrCtx`
+* feat: add New function to `ErrCtx`
 * feat: IsRootCause and RootCause are taking in account `ErrCtx` underlying errors
+* feat: RootCtxOrFallback retrieves the deepest context from wrapped errors.
 
 ## v2.0.0